Refactoring
split Core.PKX into smaller classes, only StringConverter is exposed for
outside namespaces
