+++ PKHeX.Core/PKM/CK3.cs
@@ -27,8 +27,8 @@ namespace PKHeX.Core
         }
         public override PKM Clone() { return new CK3(Data); }
 
-        public override string GetString(int Offset, int Count) => PKX.GetBEString3(Data, Offset, Count);
-        public override byte[] SetString(string value, int maxLength) => PKX.SetBEString3(value, maxLength);
+        public override string GetString(int Offset, int Count) => StringConverter.GetBEString3(Data, Offset, Count);
+        public override byte[] SetString(string value, int maxLength) => StringConverter.SetBEString3(value, maxLength);
 
         // Trash Bytes
         public override byte[] Nickname_Trash { get => GetData(0x2E, 20); set { if (value?.Length == 20) value.CopyTo(Data, 0x2E); } }
@@ -52,7 +52,7 @@ namespace PKHeX.Core
         public override ushort Checksum { get => SaveUtil.CRC16_CCITT(Data); set { } } // totally false, just a way to get a 'random' ident for the pkm.
         public override bool ChecksumValid => Valid;
 
-        public override int Species { get => PKX.GetG4Species(BigEndian.ToUInt16(Data, 0x00)); set => BigEndian.GetBytes((ushort)PKX.GetG3Species(value)).CopyTo(Data, 0x00); }
+        public override int Species { get => SpeciesConverter.GetG4Species(BigEndian.ToUInt16(Data, 0x00)); set => BigEndian.GetBytes((ushort)SpeciesConverter.GetG3Species(value)).CopyTo(Data, 0x00); }
         // 02-04 unused
         public override uint PID { get => BigEndian.ToUInt32(Data, 0x04); set => BigEndian.GetBytes(value).CopyTo(Data, 0x04); }
         public override int Version { get => SaveUtil.GetG3VersionID(Data[0x08]); set => Data[0x08] = (byte)SaveUtil.GetCXDVersionID(value); }
@@ -88,7 +88,7 @@ namespace PKHeX.Core
         public override int Move4_PP { get => Data[0x86]; set => Data[0x86] = (byte)value; }
         public override int Move4_PPUps { get => Data[0x87]; set => Data[0x87] = (byte)value; }
 
-        public override int SpriteItem => PKX.GetG4Item((ushort)HeldItem);
+        public override int SpriteItem => ItemConverter.GetG4Item((ushort)HeldItem);
         public override int HeldItem { get => BigEndian.ToUInt16(Data, 0x88); set => BigEndian.GetBytes((ushort)value).CopyTo(Data, 0x88); }
 
         // More party stats
