+++ PKHeX.Core/PKM/PK1.cs
@@ -20,8 +20,8 @@ namespace PKHeX.Core
         internal const int STRLEN_U = 11;
         private int StringLength => Japanese ? STRLEN_J : STRLEN_U;
 
-        public override string GetString(int Offset, int Count) => PKX.GetString1(Data, Offset, Count, Japanese);
-        public override byte[] SetString(string value, int maxLength) => PKX.SetString1(value, maxLength, Japanese);
+        public override string GetString(int Offset, int Count) => StringConverter.GetString1(Data, Offset, Count, Japanese);
+        public override byte[] SetString(string value, int maxLength) => StringConverter.SetString1(value, maxLength, Japanese);
 
         // Trash Bytes
         public override byte[] Nickname_Trash { get => nick; set { if (value?.Length == nick.Length) nick = value; } }
@@ -53,7 +53,7 @@ namespace PKHeX.Core
         }
         public override string Nickname
         {
-            get => PKX.GetString1(nick, 0, nick.Length, Japanese);
+            get => StringConverter.GetString1(nick, 0, nick.Length, Japanese);
             set
             {
                 byte[] strdata = SetString(value, StringLength);
@@ -71,7 +71,7 @@ namespace PKHeX.Core
 
         public override string OT_Name
         {
-            get => PKX.GetString1(otname, 0, otname.Length, Japanese);
+            get => StringConverter.GetString1(otname, 0, otname.Length, Japanese);
             set
             {
                 byte[] strdata = SetString(value, StringLength);
@@ -131,10 +131,10 @@ namespace PKHeX.Core
         #region Stored Attributes
         public override int Species
         {
-            get => PKX.GetG1Species(Data[0]);
+            get => SpeciesConverter.GetG1Species(Data[0]);
             set
             {
-                Data[0] = (byte)PKX.SetG1Species(value);
+                Data[0] = (byte)SpeciesConverter.SetG1Species(value);
 
                 // Before updating catch rate, check if non-standard
                 if (!CatchRateIsItem)
@@ -348,7 +348,7 @@ namespace PKHeX.Core
                 Move4_PP = Move4_PP,
                 Met_Location = 30013, // "Kanto region", hardcoded.
                 Gender = PersonalTable.SM[Species].RandomGender,
-                OT_Name = PKX.GetG1ConvertedString(otname, Japanese),
+                OT_Name = StringConverter.GetG1ConvertedString(otname, Japanese),
                 IsNicknamed = false,
 
                 Country = PKMConverter.Country,
@@ -364,7 +364,7 @@ namespace PKHeX.Core
             pk7.Nickname = PKX.GetSpeciesNameGeneration(pk7.Species, pk7.Language, pk7.Format);
             if (otname[0] == 0x5D) // Ingame Trade
             {
-                var s = PKX.GetG1Char(0x5D, Japanese);
+                var s = StringConverter.GetG1Char(0x5D, Japanese);
                 pk7.OT_Name = s.Substring(0, 1) + s.Substring(1).ToLower();
             }
             pk7.OT_Friendship = pk7.HT_Friendship = PersonalTable.SM[Species].BaseFriendship;
@@ -391,7 +391,7 @@ namespace PKHeX.Core
             else if (IsNicknamedBank)
             {
                 pk7.IsNicknamed = true;
-                pk7.Nickname = PKX.GetG1ConvertedString(nick, Japanese);
+                pk7.Nickname = StringConverter.GetG1ConvertedString(nick, Japanese);
             }
             
             pk7.TradeMemory(Bank:true); // oh no, memories on gen7 pkm
@@ -512,7 +512,7 @@ namespace PKHeX.Core
                 Count = Capacity;
             for (int i = 0; i < Count; i++)
             {
-                Data[1 + i] = (byte)PKX.SetG1Species(Pokemon[i].Species);
+                Data[1 + i] = (byte)SpeciesConverter.SetG1Species(Pokemon[i].Species);
                 Array.Copy(Pokemon[i].Data, 0, Data, 2 + Capacity + Entry_Size * i, Entry_Size);
                 Array.Copy(Pokemon[i].OT_Name_Raw, 0, Data, 2 + Capacity + Capacity * Entry_Size + StringLength * i, StringLength);
                 Array.Copy(Pokemon[i].Nickname_Raw, 0, Data, 2 + Capacity + Capacity * Entry_Size + StringLength * Capacity + StringLength * i, StringLength);
