+++ PKHeX.Core/PKM/PK5.cs
@@ -24,8 +24,8 @@ namespace PKHeX.Core
         }
         public override PKM Clone() { return new PK5(Data); }
 
-        public override string GetString(int Offset, int Count) => PKX.GetString5(Data, Offset, Count);
-        public override byte[] SetString(string value, int maxLength) => PKX.SetString5(value, maxLength);
+        public override string GetString(int Offset, int Count) => StringConverter.GetString5(Data, Offset, Count);
+        public override byte[] SetString(string value, int maxLength) => StringConverter.SetString5(value, maxLength);
 
         // Trash Bytes
         public override byte[] Nickname_Trash { get => GetData(0x48, 22); set { if (value?.Length == 22) value.CopyTo(Data, 0x48); } }
