+++ PKHeX.Core/Saves/SAV1.cs
@@ -483,12 +483,12 @@ namespace PKHeX.Core
             return (Data[PokedexCaughtOffset + ofs] & bitval) != 0;
         }
 
-        public override string GetString(int Offset, int Count) => PKX.GetString1(Data, Offset, Count, Japanese);
+        public override string GetString(int Offset, int Count) => StringConverter.GetString1(Data, Offset, Count, Japanese);
         public override byte[] SetString(string value, int maxLength, int PadToSize = 0, ushort PadWith = 0)
         {
             if (PadToSize == 0)
                 PadToSize = maxLength + 1;
-            return PKX.SetString1(value, maxLength, Japanese, PadToSize, PadWith);
+            return StringConverter.SetString1(value, maxLength, Japanese, PadToSize, PadWith);
         }
     }
 }
