+++ PKHeX.Core/Saves/SAV3RSBox.cs
@@ -202,12 +202,12 @@ namespace PKHeX.Core
             Edited = true;
         }
 
-        public override string GetString(int Offset, int Count) => PKX.GetString3(Data, Offset, Count, Japanese);
+        public override string GetString(int Offset, int Count) => StringConverter.GetString3(Data, Offset, Count, Japanese);
         public override byte[] SetString(string value, int maxLength, int PadToSize = 0, ushort PadWith = 0)
         {
             if (PadToSize == 0)
                 PadToSize = maxLength + 1;
-            return PKX.SetString3(value, maxLength, Japanese, PadToSize, PadWith);
+            return StringConverter.SetString3(value, maxLength, Japanese, PadToSize, PadWith);
         }
     }
 }
