+++ PKHeX.Core/Saves/SAV4BR.cs
@@ -248,12 +248,12 @@ namespace PKHeX.Core
             }
         }
 
-        public override string GetString(int Offset, int Count) => PKX.GetBEString4(Data, Offset, Count);
+        public override string GetString(int Offset, int Count) => StringConverter.GetBEString4(Data, Offset, Count);
         public override byte[] SetString(string value, int maxLength, int PadToSize = 0, ushort PadWith = 0)
         {
             if (PadToSize == 0)
                 PadToSize = maxLength + 1;
-            return PKX.SetBEString4(value, maxLength, PadToSize, PadWith);
+            return StringConverter.SetBEString4(value, maxLength, PadToSize, PadWith);
         }
     }
 }
