+++ PKHeX.WinForms/Subforms/PKM Editors/Text.cs
@@ -128,7 +128,7 @@ namespace PKHeX.WinForms
             int index = Bytes.IndexOf(nud);
             Raw[index] = (byte)nud.Value;
 
-            string str = PKX.GetString(Raw, SAV.Generation, SAV.Japanese, bigendian, Raw.Length);
+            string str = StringConverter.GetString(Raw, SAV.Generation, SAV.Japanese, bigendian, Raw.Length);
             TB_Text.Text = str;
             editing = false;
         }
@@ -138,7 +138,7 @@ namespace PKHeX.WinForms
                 return;
             editing = true;
             // build bytes
-            byte[] data = PKX.SetString(TB_Text.Text, SAV.Generation, SAV.Japanese, bigendian, Raw.Length, SAV.Language);
+            byte[] data = StringConverter.SetString(TB_Text.Text, SAV.Generation, SAV.Japanese, bigendian, Raw.Length, SAV.Language);
             Array.Copy(data, Raw, Math.Min(data.Length, Raw.Length));
             for (int i = 0; i < Raw.Length; i++)
                 Bytes[i].Value = Raw[i];
@@ -152,8 +152,8 @@ namespace PKHeX.WinForms
             if (species == "") // no result
                 species = CB_Species.Text;
 
-            byte[] current = PKX.SetString(TB_Text.Text, SAV.Generation, SAV.Japanese, bigendian, Raw.Length, SAV.Language);
-            byte[] data = PKX.SetString(species, SAV.Generation, SAV.Japanese, bigendian, Raw.Length, SAV.Language);
+            byte[] current = StringConverter.SetString(TB_Text.Text, SAV.Generation, SAV.Japanese, bigendian, Raw.Length, SAV.Language);
+            byte[] data = StringConverter.SetString(species, SAV.Generation, SAV.Japanese, bigendian, Raw.Length, SAV.Language);
             if (data.Length <= current.Length)
             {
                 WinFormsUtil.Alert("Trash byte layer is hidden by current text.",
@@ -170,7 +170,7 @@ namespace PKHeX.WinForms
         }
         private void B_ClearTrash_Click(object sender, EventArgs e)
         {
-            byte[] current = PKX.SetString(TB_Text.Text, SAV.Generation, SAV.Japanese, bigendian, Raw.Length, SAV.Language);
+            byte[] current = StringConverter.SetString(TB_Text.Text, SAV.Generation, SAV.Japanese, bigendian, Raw.Length, SAV.Language);
             for (int i = current.Length; i < Bytes.Count; i++)
                 Bytes[i].Value = 0;
         }
